# src/swagger/spec.json
{"openapi":"3.1.0","info":{"title":"ExpressStarter API","version":"1.0.0","description":"API for ExpressStarter"},"consumes":["application/json","multipart/form-data","application/x-www-form-urlencoded"],"externalDocs":{"url":"/static/docs/","description":"Find field info here"},"paths":{"/api/v1/auth/token-validity":{"get":{"summary":"Fetch auth","tags":["auth"],"security":[{"bearerAuth":[]}],"responses":{"200":{"description":"auths fetched successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}}},"parameters":[{"in":"query","name":"query","required":false,"schema":{"type":"object"}}]}},"/api/v1/user/profile":{"get":{"summary":"Fetch user","tags":["user"],"security":[{"bearerAuth":[]}],"responses":{"200":{"description":"users fetched successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}}},"parameters":[{"in":"query","name":"query","required":false,"schema":{"type":"object"}}]},"put":{"summary":"Update user","tags":["user"],"security":[{"bearerAuth":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GeneralBody"}}}},"responses":{"200":{"description":"user updated successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}}}}},"/api/v1/user/profiles":{"get":{"summary":"Fetch user","tags":["user"],"security":[{"bearerAuth":[]}],"responses":{"200":{"description":"users fetched successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}}},"parameters":[{"in":"query","name":"query","required":false,"schema":{"type":"object"}}]}},"/api/v1/user":{"get":{"summary":"Fetch user","tags":["user"],"security":[{"bearerAuth":[]}],"responses":{"200":{"description":"users fetched successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}}},"parameters":[{"in":"query","name":"query","required":false,"schema":{"type":"object"}}]},"put":{"summary":"Update user","tags":["user"],"security":[{"bearerAuth":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GeneralBody"}}}},"responses":{"200":{"description":"user updated successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}}}}},"/api/v1/user/update-status":{"put":{"summary":"Update user","tags":["user"],"security":[{"bearerAuth":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GeneralBody"}}}},"responses":{"200":{"description":"user updated successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}}}}},"/api/v1/user/{id}":{"delete":{"summary":"Delete user","tags":["user"],"security":[{"bearerAuth":[]}],"responses":{"200":{"description":"user deleted successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}}},"parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"string"}}]}},"/api/v1/user/all/data":{"delete":{"summary":"Delete user","tags":["user"],"security":[{"bearerAuth":[]}],"responses":{"200":{"description":"user deleted successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}}}}},"/api/v1/auth/login":{"post":{"summary":"Create a new auth","tags":["auth"],"security":[{"bearerAuth":[]}],"requestBody":{"required":false,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GeneralBody"}}}},"responses":{"200":{"description":"auth created successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}}}}},"/api/v1/file":{"post":{"summary":"Create a new file","tags":["file"],"security":[{"bearerAuth":[]}],"requestBody":{"required":false,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GeneralBody"}}}},"responses":{"200":{"description":"file created successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}}}},"get":{"summary":"Fetch file","tags":["file"],"security":[{"bearerAuth":[]}],"responses":{"200":{"description":"files fetched successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}}},"parameters":[{"in":"query","name":"query","required":false,"schema":{"type":"object"}}]}},"/api/v1/file/{id}":{"get":{"summary":"Fetch file","tags":["file"],"security":[{"bearerAuth":[]}],"responses":{"200":{"description":"files fetched successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}}},"parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"string"}},{"in":"query","name":"query","required":false,"schema":{"type":"object"}}]},"put":{"summary":"Update file","tags":["file"],"security":[{"bearerAuth":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GeneralBody"}}}},"responses":{"200":{"description":"file updated successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}}}},"delete":{"summary":"Delete file","tags":["file"],"security":[{"bearerAuth":[]}],"responses":{"200":{"description":"file deleted successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}}}}},"/api/v1/single/{model}/{id}":{"get":{"summary":"Fetch single","tags":["single"],"security":[{"bearerAuth":[]}],"responses":{"200":{"description":"singles fetched successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}}},"parameters":[{"in":"path","name":"model","required":true,"schema":{"type":"string"}},{"in":"path","name":"id","required":true,"schema":{"type":"string"}},{"in":"query","name":"query","required":false,"schema":{"type":"object"}}]}},"/api/v1/auth/register":{"post":{"summary":"Create a new auth","tags":["auth"],"security":[{"bearerAuth":[]}],"requestBody":{"required":false,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GeneralBody"}}}},"responses":{"200":{"description":"auth created successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}}}}},"/api/v1/auth/change-password":{"post":{"summary":"Create a new auth","tags":["auth"],"security":[{"bearerAuth":[]}],"requestBody":{"required":false,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GeneralBody"}}}},"responses":{"200":{"description":"auth created successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}}}}},"/api/v1/auth/verify-email-account":{"post":{"summary":"Create a new auth","tags":["auth"],"security":[{"bearerAuth":[]}],"requestBody":{"required":false,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GeneralBody"}}}},"responses":{"200":{"description":"auth created successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}}}}},"/api/v1/auth/request-email-verification":{"post":{"summary":"Create a new auth","tags":["auth"],"security":[{"bearerAuth":[]}],"requestBody":{"required":false,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GeneralBody"}}}},"responses":{"200":{"description":"auth created successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}}}}},"/api/v1/auth/request-phone-verification":{"post":{"summary":"Create a new auth","tags":["auth"],"security":[{"bearerAuth":[]}],"requestBody":{"required":false,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GeneralBody"}}}},"responses":{"200":{"description":"auth created successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}}}}},"/api/v1/auth/verify-phone":{"post":{"summary":"Create a new auth","tags":["auth"],"security":[{"bearerAuth":[]}],"requestBody":{"required":false,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GeneralBody"}}}},"responses":{"200":{"description":"auth created successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}}}}},"/api/v1/auth/request-reset-password":{"post":{"summary":"Create a new auth","tags":["auth"],"security":[{"bearerAuth":[]}],"requestBody":{"required":false,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GeneralBody"}}}},"responses":{"200":{"description":"auth created successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}}}}},"/api/v1/auth/reset-password":{"post":{"summary":"Create a new auth","tags":["auth"],"security":[{"bearerAuth":[]}],"requestBody":{"required":false,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GeneralBody"}}}},"responses":{"200":{"description":"auth created successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}}}}}},"components":{"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}},"schemas":{"GeneralBody":{"type":"object"},"QueryParams":{"type":"object"},"ApiResponse":{"type":"object","properties":{"message":{"type":"string"},"data":{"type":"object"}}},"RegisterDto":{"type":"object","properties":{"email":{"type":"any"},"password":{"type":"any"},"firstName":{"type":"any"},"lastName":{"type":"any"},"type":{"type":"any"},"phone":{"type":"any"}}},"LoginDto":{"type":"object","properties":{"username":{"type":"any"},"password":{"type":"any"}}},"ChangePasswordDto":{"type":"object","properties":{"oldPassword":{"type":"any"},"newPassword":{"type":"any"}}},"ResetPasswordDto":{"type":"object","properties":{"token":{"type":"any"},"newPassword":{"type":"any"}}},"ForgotPasswordDto":{"type":"object","properties":{"email":{"type":"any"}}},"VerifyEmailResendDto":{"type":"object","properties":{"email":{"type":"any"}}},"VerifyToken":{"type":"object","properties":{"token":{"type":"any"}}},"RequestPhoneVerificationDto":{"type":"object","properties":{"phone":{"type":"any"}}},"CreateFileDto":{"type":"object","properties":{"url":{"type":"any"},"name":{"type":"any"},"type":{"type":"any"},"size":{"type":"any"},"refType":{"type":"any"},"fieldname":{"type":"any"},"ref":{"type":"any"},"format":{"type":"any"},"createdBy":{"type":"any"},"metadata":{"type":"any"}}},"UpdateFileDto":{"type":"object","properties":{}},"CreateUserDto":{"type":"object","properties":{}},"UpdateUserDto":{"type":"object","properties":{"email":{"type":"any"},"phone":{"type":"any"}}},"UpdateUserStatusDto":{"type":"object","properties":{"id":{"type":"any"},"status":{"type":"any"},"description":{"type":"any"}}},"updateEmailDto":{"type":"object","properties":{"email":{"type":"any"}}}}},"tags":[]}